Modify Storer interface and add explanation
Initially, it was thought that a Storer should be safe for concurrent use by multiple goroutines.

As safe as it might seem, it ends up introducing non-determinism in the exact content of a storage datastructure for a program that does share a common store across mutliple goroutines.

What we want is for the storage to be mutable within a single goroutine.
But we now also discourage for goroutines to have shared local storage.

It implies a change in implementation of the Spawn method of a context as the store is now cloned(deep copy).
The presence of an item in the storer shall be independent from the scheduling of goroutines.