Change comment re. execution.Context Storer usage.
The Storer field of the execution.Context object is used for goroutine local storage of data.
The emphasis on locality means that it is not expected for a Storer to be shared across multiple goroutines.
However, it may be merely deep copied as it is done each time a subtask is created via the launch of a new goroutine using a "spawned context".